requirements.txt
opencv-python==4.6.0.66
pyzmq==23.2.0
